fix(terrain helper): fix terrain helper hook
